+++ R/simple_logging.R
@@ -439,6 +439,10 @@ find_target <- function(
 ){
   assert(is_scalar_character(name))
 
+  if (is_scalar_character(x)) {
+    x <- get_logger(x)
+  }
+
   if (inherits(x, "Appender")){
     if (name %in% names(x)){
       return(x)