+++ tests/testthat/test_Appender.R
@@ -1,8 +1,6 @@
 context("Appender")
 
 
-
-
 x <- LogEvent$new(
   logger = Logger$new("dummy"),
   level = 200L,