+++ tests/testthat/test_utils-rd.R
@@ -1,0 +1,12 @@
+context("utils-rd")
+
+
+test_that("utils-rd works as expected", {
+
+  r6_usage(list(
+    AppenderFileRotating,
+    AppenderFileRotatingDate,
+    AppenderFileRotatingTime
+  ))
+
+})