+++ NEWS.md
@@ -13,7 +13,8 @@
 * `with_log_level()` and `with_log_value()` now accept logger names as well as 
   Logger objects as the `logger` argument
 * `get_logger_glue()` now works as intended
-  
+* `FATAL()`, `ERROR()`,... are now deprecated and will be removed from future
+  versions. Use lgr$fatal(), lgr$error(), ... instead.
 
 
 # lgr 0.2.1 (2019-03-25)