+++ NEWS.md
@@ -1,4 +1,12 @@
-# lgr 0.2.1.9000
+# lgr 0.2.2.9000
+  * `AppenderFileRotating`, `AppenderFileRotatingDate` and 
+    `AppenderFileRotatingTime` for rotating file appenders via the 
+    [rotor](https://github.com/s-fleck/rotor).
+  * functions like `show_log()`, `show_data()`,... now accept logger names
+    as `target`.
+
+
+# lgr 0.2.2 (unreleased)
 
 * The root logger can now be configured via `options()` and/or environment 
   variables (see `?lgr`)